+++ src/components/mod.rs
@@ -1,1 +1,2 @@
 pub mod comment;
+pub mod command;
