+++ src/main.rs
@@ -13,7 +13,8 @@ mod grammar;
 mod components;
 
 use crate::components::comment::CommentNode;
-use crate::format::{Format, text};
+use crate::components::command::CommandNode;
+use crate::format::{Format, text, Measure};
 use crate::state::{Config, BlockState};
 use crate::grammar::{HyprlangParser, Rule};
 use pest::iterators::Pair;
@@ -48,12 +49,6 @@ struct Block {
     nodes: Vec<Node>,
 
     config: Config,
-}
-
-trait Measure {
-    fn as_lhs(&self) -> Option<String>;
-    fn as_rhs(&self) -> Option<String>;
-    fn as_mid(&self) -> Option<String>;
 }
 
 impl Block {
@@ -104,58 +99,6 @@ impl fmt::Display for Block {
     }
 }
 
-#[derive(PartialEq)]
-struct CommandNode {
-    comment: Option<String>,
-    ident: String,
-    parts: Vec<String>,
-}
-
-impl Format for CommandNode {
-    fn format(&self, _config: Config, state: &BlockState) -> Result<String, fmt::Error> {
-        let lhs_pad_right = state.lhs_max_length;
-
-        let lhs = self.as_lhs().expect("infallible");
-        let mid = self.as_mid().expect("infallible");
-        let rhs = self.as_rhs().expect("infallible");
-
-        let mut s = String::new();
-        write!(s, "{lhs:lhs_pad_right$}{mid}{rhs}")?;
-
-        if let Some(c) = &self.comment {
-            let comment_gap = state.max_length - s.as_str().len();
-            write!(s, " {empty:>comment_gap$}{c}", empty = "")?;
-        }
-        Ok(s)
-    }
-}
-
-impl CommandNode {
-    fn new(tag: &Pair<Rule>) -> Self {
-        let mut ident = None;
-        let mut parts = Vec::new();
-        let mut comment = None;
-
-        for part in tag.clone().into_inner() {
-            match part.as_rule() {
-                Rule::command_ident => {
-                    ident = Some(text(&part));
-                }
-                Rule::command_expression | Rule::command_rule => parts.push(text(&part)),
-                Rule::comment => {
-                    comment = Some(text(&part));
-                }
-                _ => unreachable!(),
-            }
-        }
-
-        CommandNode {
-            comment,
-            ident: ident.expect("command must have an ident"),
-            parts,
-        }
-    }
-}
 #[derive(PartialEq)]
 struct VariableAssignmentNode {
     comment: Option<String>,
@@ -321,30 +264,6 @@ impl Measure for VariableAssignmentNode {
     }
     fn as_rhs(&self) -> Option<String> {
         Some(self.expression.to_string())
-    }
-    fn as_mid(&self) -> Option<String> {
-        let has_lhs = self.as_lhs().map(|side| !side.is_empty());
-        let has_rhs = self.as_rhs().map(|side| !side.is_empty());
-        match (has_lhs, has_rhs) {
-            (Some(l), Some(r)) if l && r => Some(String::from(" = ")),
-            (Some(l), Some(r)) if l && !r => Some(String::from(" =")),
-            _ => None,
-        }
-    }
-}
-
-impl Measure for CommandNode {
-    fn as_lhs(&self) -> Option<String> {
-        Some(self.ident.to_string())
-    }
-    fn as_rhs(&self) -> Option<String> {
-        Some(
-            self.parts
-                .iter()
-                .map(std::string::ToString::to_string)
-                .collect::<Vec<_>>()
-                .join(" "),
-        )
     }
     fn as_mid(&self) -> Option<String> {
         let has_lhs = self.as_lhs().map(|side| !side.is_empty());
